requirements.txt
streamlit
google-generativeai
pillow
pandas
plotly